fix: import versions correctly
Previously, we weren't correctly importing versioned documents. I
incorrectly assumed we could just use the `version` property, and that
we could ignore `links`.

This updates the importer to correctly create and update documents.

Effectively, it loops over all of a document's versions, creating or
updating them when all of its links are available.